pyproject.toml
[project]
name = "tms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "pydantic-settings>=2.10.1",
    "pyright>=1.1.404",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
]
